// src/lib/schema.ts
export const db_add_faq_sc = z.object({
 question: z.string().nonempty("Question is required"),
 answer: z.string().nonempty("Answer is required"),
});
export const db_edit_faq_sc = z.object({
 question: z.string().optional(),
 answer: z.string().optional(),
});